+++ src/product/tool-definitions.js
@@ -58,6 +58,7 @@ import { listDataToolDefinitions } from './definitions/list-data-tools.js';
 import { dataFormatToolDefinitions } from './definitions/data-format-tools.js';
 import { fileUtilityToolDefinitions } from './definitions/file-utility-tools.js';
 import { archiveFileToolDefinitions } from './definitions/archive-file-tools.js';
+import { webContentToolDefinitions } from './definitions/web-content-tools.js';
 import { financeDefinitions } from './definitions/finance.js';
 import { healthDefinitions } from './definitions/health.js';
 import { mathDefinitions } from './definitions/math.js';
@@ -1085,6 +1086,7 @@ const toolDefinitions = Object.freeze({
     ...dataFormatToolDefinitions,
     ...fileUtilityToolDefinitions,
     ...archiveFileToolDefinitions,
+    ...webContentToolDefinitions,
 });
 
 function getToolDefinition(id) {